✨ Introduced admin utils
